huge design update + float to bigdecimal

+++ src/main/java/by/bsu/d0mpi/UP_PostGallery/command/action/DeletePostAction.java
@@ -8,6 +8,8 @@ import by.bsu.d0mpi.UP_PostGallery.service.PostService;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
+import javax.servlet.http.HttpSession;
+
 public class DeletePostAction implements Command {
     private static final Logger logger = LogManager.getLogger();
     private static volatile DeletePostAction instance;
@@ -35,7 +37,21 @@ public class DeletePostAction implements Command {
 
     @Override
     public CommandResponse execute(CommandRequest request) {
-        Integer postId = Integer.valueOf(request.getParameter("post_id"));
+        Integer postId;
+        HttpSession session = request.getCurrentSession().orElse(null);
+        try {
+            postId = Integer.valueOf(request.getParameter("post_id"));
+        } catch (NumberFormatException e) {
+            e.printStackTrace();
+            return redirectHomePage;
+        }
+        if (session == null ||
+                postService.findEntityById(postId) == null ||
+                !postService.doesPostBelongsToAuthor(postId, (String) session.getAttribute("user_name")) &&
+                        !(session.getAttribute("current_role").toString().equals("ADMIN") ||
+                                session.getAttribute("current_role").toString().equals("MODERATOR"))) {
+            return redirectHomePage;
+        }
         postService.delete(postId);
         return redirectHomePage;
     }
